First and Final Commit
This was pretty easy. A lot of it was stuff we had gone over before or stuff that I had picked up on while doing the unit projects. There was a try statement in the code that I didn't get really the effects with the try statements could have been done with min and max in the html code.